refactor(api): PR-6.5 commit 1 — projects/versions CRUD 16 routes
server.py 2382 → 2122 行（−260），projects/versions 域第一刀。

  api/routers/projects/__init__.py       说明文档（5 子文件计划）
  api/routers/projects/_shared.py        7 helper：_project_payload /
                                         _publish_project_state / _publish_version_state /
                                         _project_err_code / _version_dir_or_404 /
                                         _version_train_dir_or_404 / _project_and_version_or_404 /
                                         _reg_dir
  api/routers/projects/crud.py    16     list/create/get/patch/delete project +
                                  routes  list/create/get/patch/delete version +
                                          activate / advance-phase / skip-phase +
                                          lora_ckpts / state_ckpts / lora_ckpts
  api/schemas/projects.py         4 BaseModel  ProjectCreate / ProjectUpdate /
                                               VersionCreate / VersionUpdate

projects/ 子包设计（同 queue/ 模式）：
- _shared.py 装 projects 域内跨 sub-router 共用 helper（不升 api/deps.py
  —— 避免非 projects router 反向耦合到 projects 域细节）
- 5 个 sub-router（crud / exports / ingestion / curation / training）按
  domain 拆，app.py 各自 include
- commit 2-5 把剩余 55 routes 渐进搬走，server.py 内 reg / version_config /
  curation handler 通过顶部 import 复用 _shared.py 中的 _version_dir_or_404 等

回归：PR-1 安全网 152 + project_endpoints 49 = **201/201 全绿**。
route snapshot 不变（160 routes）。
