+++ studio/api/app.py
@@ -34,6 +34,7 @@ from .routers import (
     tagger,
     upscalers,
 )
+from .routers.projects import crud as projects_crud
 from .routers.queue import io as queue_io_router
 from .routers.queue import lifecycle as queue_lifecycle
 from .routers.queue import outputs as queue_outputs
@@ -71,3 +72,5 @@ app.include_router(generate.router)
 app.include_router(queue_io_router.router)
 app.include_router(queue_lifecycle.router)
 app.include_router(queue_outputs.router)
+# PR-6.5 commit 1: projects/versions CRUD 子包第一刀（16 routes）
+app.include_router(projects_crud.router)
